requirements.txt
pandas
numpy
scikit-learn
nltk
matplotlib
seaborn
fastapi
uvicorn
python-multipart
joblib
transformers
torch
ipykernel
requests
beautifulsoup4
newspaper3k
lxml_html_clean
